+++ ElvUI_EltreumUI/Core/Options/Changelog.lua
@@ -14,28 +14,35 @@ function ElvUI_EltreumUI:Changelog()
 	ElvUI_EltreumUI.Options.args.changelog.args.added = E.Libs.ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = E.Libs.ACH:Description([[
-Added an option to disable Borders on Raid Unitframe Auras
+Added an option to disable the mouseover of the Damage Meter
+Added an option to disable the coloring of Rare Nameplates
+Added an option to disable Mirror Timer skins
+Added [name:eltruism:gradientreverse] tag
 ]], 3, "small", nil, nil, nil, nil, "full")
 
 	--updated
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = E.Libs.ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = E.Libs.ACH:Description([[
-Updated Gradient Nameplates to fix issue with flashing threat colors
-Updated Portraits naming scheme
-Updated Damage Meter skin
-Updated Nameplate CVars
+Updated Damage Meter skin to add Acidweb's fix for the clamping
+Updated some Datatext options that are no longer used
+Updated the nameplate backdrop remover to work again
+Updated several tags to solve issues with secrets
+Updated Retail talent scale (thanks Sash)
+Updated German locale by Dlarge
+Updated Windtools profile
 --]], 5, "small", nil, nil, nil, nil, "full")
 
 	--fixed
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = E.Libs.ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = E.Libs.ACH:Description([[
-Fixed an issue with Borders and changes in ElvUI debuff types
-Fixed an issue with the Burning Crusade Profession skin
-Fixed an issue with the Alternate Power Bar and secrets
-Fixed issues with secrets related to Tooltip borders
-Fixed an issue with the Sockets in Burning Crusade
-Fixed an issue when changing Portrait settings
+Fixed an issue where the Holy Paladin spec icon had a Devourer Demon Hunter overlayed
+Fixed several issues related to PVP instances after the hotfix made them secret
+Fixed an issue where the Leave Vehicle Button no longer had the Texture
+Fixed an issue where some Portrait options were not applying
+Fixed an issue related to using transparent Power
+Fixed an issue with Borders and secrets
+Fixed issues related to chat in Retail
 ]], 7, "small", nil, nil, nil, nil, "full")
 end