+++ Changelog.md
@@ -1,6 +1,26 @@
 ### 🌌 Next Version
 The future is uncertain
 ---
+### 5.0.2
+Added an option to disable the mouseover of the Damage Meter  
+Added an option to disable the coloring of Rare Nameplates  
+Added an option to disable Mirror Timer skins  
+Added [name:eltruism:gradientreverse] tag  
+Updated Damage Meter skin to add Acidweb's fix for the clamping  
+Updated some Datatext options that are no longer used  
+Updated the nameplate backdrop remover to work again  
+Updated several tags to solve issues with secrets  
+Updated Retail talent scale (thanks Sash)  
+Updated German locale by Dlarge  
+Updated Windtools profile  
+Fixed an issue where the Holy Paladin spec icon had a Devourer Demon Hunter overlayed  
+Fixed several issues related to PVP instances after the hotfix made them secret  
+Fixed an issue where the Leave Vehicle Button no longer had the Texture  
+Fixed an issue where some Portrait options were not applying  
+Fixed an issue related to using transparent Power  
+Fixed an issue with Borders and secrets  
+Fixed issues related to chat in Retail  
+---
 ### 5.0.1
 Added an option to disable Borders on Raid Unitframe Auras  
 Updated Gradient Nameplates to fix issue with flashing threat colors